add error state ui

+++ app/src/main/java/ru/sigarev/whattowear/domain/models/DayTemperature.kt
@@ -1,8 +1,8 @@
 package ru.sigarev.whattowear.domain.models
 
 data class DayTemperature(
-    val day: Double,
-    val eve: Double,
-    val morn: Double,
-    val night: Double
+    val day: Int,
+    val eve: Int,
+    val morn: Int,
+    val night: Int
 )
