+++ app/src/main/java/ru/sigarev/whattowear/domain/models/LocationWithTemperature.kt
@@ -2,5 +2,5 @@ package ru.sigarev.whattowear.domain.models
 
 data class LocationWithTemperature(
     val location: Location,
-    val currentTemperature: Float
+    val currentTemperature: Int
 )
